Add tests for eval.py
